+++ Source/WinCopies.Util.Shared/Exceptions/InvalidEnumArgumentException.cs
@@ -18,6 +18,7 @@
 using System;
 using System.Runtime.Serialization;
 using System.Security;
+
 using WinCopies.Collections;
 
 #if WinCopies2